requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-environ==0.11.2
django-cors-headers==4.3.1